refactor(claude-hooks): move the cold-start wait into hook-io, name it for what it waits on
The three primitives wait on any lazily-loaded dependency, not on the control
plane specifically — scan-invisible-chars uses them for agent-sanitizer. Keeping
them in lib/control-plane made the SessionStart scanner import a module whose
body awaits agent-control-plane-core, a package that hook never touches, so the
scan paid a package load for three functions. hook-io already owns lazyImport
and markerIsTrusted, which these build on and which pull in nothing.
